Smart slicing - Memory refactor
1. Utmos functional tests are now passing again such that everything `--lowmem` works.
2. I've renamed the parameter `--chunk-length` to `--buffer`
3. There's a new parameter `--maxmem` which is the maximum memory utmos tries to use when
doing the summations. By default it is 2 (GB).

Previously, the lowmem functionality was IO bound and very slow. The
problem is we were calling on HDF5 to read every chunk. This change
tries to reduce the number of calls to HDF5 by doing chunk-size to
maximum memory guesstimation to minimize the number of slice/sum
operations on the h5py Dataset. Now, we only make a call to the hdf5
file on ~ maxmem / (size of 1 chunk) number of rows. Additionally, we
'short-circuit' the lowmem/hdf5 workflow by identifying when the amount
of data remaining after an iteration is small enough to fit in the
~maxmem. At that point we read everything into memory and abandon the
hdf5 intermediate file strategy.